Corrected fullRotation type in SteeringWheel

+++ RobotSteeringWheel.h
@@ -16,7 +16,7 @@ class RobotSteeringWheel
  private:
 	 Mistumi95SP9PUnipolar stepper;
 	 static const byte left = 3, right = 2, straight = 1;
-	 byte fullRotation;
+	 float fullRotation;
 	 byte direction;
  public:
 	RobotSteeringWheel();
